% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m_data}
\alias{m_data}
\title{Simulated voting data (multinomial)}
\format{
a matrix, rows are voters (I = 100) and cols are bills (J = 15).
}
\usage{
m_data
}
\description{
\code{m_data} is simulated voting dataset
}
\keyword{datasets}
