% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m_data_dyn}
\alias{m_data_dyn}
\title{Simulated voting long-format data (dynamic multinomial)}
\format{
a tbl_df with 100 voters, 120 bills and 10 sessions. Observations are 11,597 (NAs are dropped) with 4 columns:
\describe{
\item{unit}{IDs of voters}
\item{time}{IDs of sessions}
\item{bill}{IDs of bills}
\item{vote}{Voting choice of a voter}
}
}
\usage{
m_data_dyn
}
\description{
\code{m_data_dyn} is simulated voting long-format dataset
}
\keyword{datasets}
