% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bill_match.R
\name{make_bill_match}
\alias{make_bill_match}
\title{Generating matched bill indicator}
\usage{
make_bill_match(data, bill_pairs)
}
\arguments{
\item{data}{a matrix object, roll-call matrix.}

\item{bill_pairs}{a data.frame, tbl_df or matrix object (any x 2), the first column contains ID (integer or numeric) of bills
which must be matched with the colunm names of \code{data} and the second column contains matched bills' ID before the bill.}
}
\description{
\code{make_bill_match} is an auxiliary function for pgIRT model and generates matched bill indicator for dynamic pgIRT models.
}
\examples{
data(m_data_dyn)
data(sim_match)
\dontrun{
mat <- make_rollcall(m_data_dyn,
                     unit_id = "unit",
                     bill_id = "bill",
                     vote_col = "vote")
mat <- clean_rollcall(mat)
bill_match <- make_bill_match(mat, sim_match)
}
}
