% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_init.R
\name{make_init}
\alias{make_init}
\title{Generating initial values for pgIRT model}
\usage{
make_init(
  data,
  model = c("bin", "bin_dyn", "multi", "multi_dyn"),
  T = NULL,
  bill_match = NULL,
  constraint = NULL
)
}
\arguments{
\item{data}{a matrix object, roll-call matrix}

\item{model}{string, one of "bin" (binary), "bin_dyn" (dynamic binary), "multi" (multinomial) or "multi_dyn" (dynamic multinomial)}

\item{T}{an integer, indicating the number of sessions. If you choose "bin_dyn" or "multi_dyn" for \code{model}, you must supply this argument.}

\item{bill_match}{(Optional). (Optional). An integer vector which indicates matched bills (the location of matched bills, starts from 0)
and is the same length as the number of bills. If a bill does not have matches, please fill NA. Using \link{make_bill_match} is strongly recommended.}

\item{constraint}{an integer or integer vector (for dynamic model, T-length), indicating the voter whose ideal point is always set positive.}
}
\description{
\code{make_init} is an auxiliary function for pgIRT model and generates optimal initial values based on roll-call matrix.
}
