% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rollcall.R
\name{make_rollcall}
\alias{make_rollcall}
\title{Generating roll-call matrix for pgIRT model}
\usage{
make_rollcall(
  dataframe,
  unit_id = NULL,
  bill_id = NULL,
  vote_col = NULL,
  drop_unanimous = FALSE
)
}
\arguments{
\item{dataframe}{data.frame, tbl_df or matrix object, dataframe (long-format, i.e. voter-bill unit) of voting data.}

\item{unit_id}{a string, column name which indicates voters. Integer or numeric column is highly recommended.}

\item{bill_id}{a string, column name which indicates bills. Only integer or numeric column is allowed.}

\item{vote_col}{a string, column name which indicates the votes. Only integer or numeric column is allowed.}

\item{drop_unanimous}{a bool, whether removing unanimous bills or not. Default is FALSE.}
}
\description{
\code{make_rolllcall} is an auxiliary function for pgIRT model and generates roll-call matrix from long-dataframe.
}
