% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgIRT.R
\name{pgIRT}
\alias{pgIRT}
\title{Polya-Gamma IRT with EM algorithm}
\usage{
pgIRT(
  data,
  model = c("bin", "bin_dyn", "multi", "multi_dyn"),
  prior = NULL,
  init = NULL,
  constraint = NULL,
  dyn_options = NULL,
  tol = 1e-06,
  maxit = 500,
  verbose = NULL,
  std = FALSE
)
}
\arguments{
\item{data}{a matrix, roll-call matrix (I x J). For binary model, the matrix is allowed to contain only 1, 0 and NA. For multinomial, 1, 2, 3 and NA is allowed.}

\item{model}{a string, one of "bin" (binary), "bin_dyn" (dynamic binary), "multi" (multinomial) or "multi_dyn" (dynamic multinomial)}

\item{prior}{a list, containing prior distribution:
\itemize{
\item \code{a0} a double (\code{bin}, \code{bin_dyn})  or 2-length vector (\code{multi}, \code{multi_dyn}), prior mean of alpha. Default is 0 (\code{bin}, \code{bin_dyn}), c(0, 0) (\code{multi}, \code{multi_dyn}).
\item \code{A0} a double (\code{bin}, \code{bin_dyn}) or 2-length vector (\code{multi}, \code{multi_dyn}), prior variance of alpha. Default is 25 (\code{bin}, \code{bin_dyn}), c(25, 25) (\code{multi}, \code{multi_dyn}).
\item \code{b0} a double (\code{bin}, \code{bin_dyn}) or 2-length vector (\code{multi}, \code{multi_dyn}), prior mean of beta Default is 0 (\code{bin}, \code{bin_dyn}), c(0, 0) (\code{multi}, \code{multi_dyn}).
\item \code{B0} a double (\code{bin}, \code{bin_dyn}) or 2-length vector (\code{multi}, \code{multi_dyn}), prior variance of beta. Default is 25 (\code{bin}, \code{bin_dyn}), c(25, 25) (\code{multi}, \code{multi_dyn}).
\item \code{theta0} a numeric vector (I length), prior mean of theta_i0 for dynamic model (\code{bin_dyn}, \code{multi_dyn}). Default is \code{rep(0, I)}.
\item \code{Delta0} a numeric vector (I length), prior variance of theta_i0 for dynamic model (\code{bin_dyn}, \code{multi_dyn}). Default is \code{rep(1, I)}.
\item \code{Delta} a double, prior evolution variance of theta_it for dynamic model (\code{bin_dyn}, \code{multi_dyn}). Default is 0.01.
}}

\item{init}{a list, containing initial values (strongly recommended to use \link{make_init}):
\itemize{
\item \code{alpha} J length vector of alpha for \code{bin} and \code{bin_dyn} model. J x 2 matrix for \code{multi} and \code{multi_dyn}.
\item \code{beta} J length vector of beta for \code{bin} and \code{bin_dyn} model. J x 2 matrix for \code{multi} and \code{multi_dyn}.
\item \code{theta} I length vector of theta for \code{bin} and \code{multi} model. I x T (sessions) matrix for \code{bin_dyn} and \code{multi_dyn}.
}}

\item{constraint}{an integer or integer vector (for dynamic model, the same length as the number of sessions), indicating the voter whose ideal point is always set positive.}

\item{dyn_options}{a list, containing the options for dynamic model. If you choose \code{bin_dyn} or \code{multi_dyn} for \code{model}, you must supply this argument. Using \link{make_dyn_options}() is strongly recommended:
\itemize{
\item \code{session_individual} a matrix, the first column contains attending sessions (starts from 0) of individuals and
the second column contains the identifiers of individuals.
\item \code{bill_session} an integer vector, indicating the sessions each bill belongs to (starts from 0).
\item \code{bill_match} (Optional). An integer vector which indicates matched bills (the location of matched bills, starts from 0)
and is the same length as the number of bills. If a bill does not have matches, please fill NA. Using \link{make_bill_match} is strongly recommended.
}}

\item{tol}{a double (< 1), convergence threshold. Default is 1e-6.}

\item{maxit}{am integer, maximum number of iterations for EM. Default is 500.}

\item{verbose}{a integer, the function prints the status every \code{verbose}. Default is NULL.}

\item{std}{a bool, whether ideal points are standardized or not. Default is FALSE.}
}
\description{
\link{pgIRT} Estimates IRT model (binary, multinomial and dynamic) by Polya-Gamma data augmentation and EM algorithm.
}
