% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgIRT_boot.R
\name{pgIRT_boot}
\alias{pgIRT_boot}
\title{Parametric bootstrap for pgIRT}
\usage{
pgIRT_boot(fit, boot = 100, verbose = NULL)
}
\arguments{
\item{fit}{pgIRT object from \code{pgIRT()}.}

\item{boot}{an integer, number of bootstrap iterations. Default is 100.}

\item{verbose}{an integer, the function prints the status every \code{verbose}. Default is NULL.}
}
\description{
\link{pgIRT_boot} implements paramtric bootstrap for pgIRT model to compute statistical uncertainty.
}
\examples{
\dontrun{
# Multinomial model (K >= 2)
data(m_data_dyn)
mat <- make_rollcall(m_data_dyn,
                     unit_id = "unit",
                     bill_id = "bill",
                     vote_col = "vote")
mat <- clean_rollcall(mat)
pgfit_mlt <- pgIRT(mat,
                   model = "default",
                   constraint = 1)
boot <- pgIRT_boot(pgfit_mlt, boot = 100, verbose = 10)
summary(boot)
}
}
