% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.pgIRT_fit.R
\name{predict.pgIRT_fit}
\alias{predict.pgIRT_fit}
\title{Predict function for \code{pgIRT_fit}}
\usage{
\method{predict}{pgIRT_fit}(object, type = "prob", ...)
}
\arguments{
\item{object}{an object class of \code{pgIRT_fit}}

\item{type}{a string, one of "prob" or "response". "prob" returns predicted probabilities of choices
and "response" returns predicted value of Y.}

\item{...}{parameters to \code{predict()}.}
}
\description{
Predict function for \code{pgIRT_fit}
}
