% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.pgIRT.R
\name{print.summary.pgIRT_boot}
\alias{print.summary.pgIRT_boot}
\title{Print function for EM boot}
\usage{
\method{print}{summary.pgIRT_boot}(object)
}
\arguments{
\item{object}{An output of \code{summary.pgIRT_boot()} function.}
}
\description{
Print function for EM boot
}
