% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_match}
\alias{sim_match}
\title{Simulated matched bills (dynamic multinomial)}
\format{
a tbl_df with 20 matched votes (40 x 2):
\describe{
\item{rc1}{IDs of bills which have a matched bill}
\item{rc2}{IDs of bills which are matched with \code{rc1}}
}
}
\usage{
sim_match
}
\description{
\code{sim_match} is simulated matched bills for dynamic estimation
}
\keyword{datasets}
