% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.pgIRT.R
\name{summary.pgIRT_boot}
\alias{summary.pgIRT_boot}
\title{Summary function for EM boot}
\usage{
\method{summary}{pgIRT_boot}(object, parameter = c("alpha", "beta", "theta"), ci = 0.95, ...)
}
\arguments{
\item{object}{an object class of \code{"pgIRT_boot"}}

\item{parameter}{a string or string vector, indicating what parameters you want to get.}

\item{ci}{a float (< 1). The function returns \code{ci} * 100 \% confidence interval.}

\item{...}{parameters to \code{summary()}.}
}
\description{
Summary function for EM boot
}
