% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.pgIRT.R
\name{summary.pgIRT_fit}
\alias{summary.pgIRT_fit}
\title{Summary function for EM}
\usage{
\method{summary}{pgIRT_fit}(object, parameter = c("alpha", "beta", "theta"), ...)
}
\arguments{
\item{object}{an object class of \code{"pgIRT_fit"}}

\item{parameter}{a string or string vector, indicating what parameters you want to get.}

\item{...}{parameters to \code{summary()}.}
}
\description{
Summary function for EM
}
