% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dyn_options.R
\name{make_dyn_options}
\alias{make_dyn_options}
\title{Generating options for dynamic pgIRT model}
\usage{
make_dyn_options(
  dataframe,
  unit_id = NULL,
  bill_id = NULL,
  time_id = NULL,
  vote_col = NULL,
  bill_match = NULL,
  drop_unanimous = FALSE
)
}
\arguments{
\item{dataframe}{data.frame, tbl_df or matrix object, dataframe (long-format, i.e. voter-bill unit) of voting data.}

\item{unit_id}{a string, column name which indicates voters. Integer or numeric column is allowed.}

\item{bill_id}{a string, column name which indicates bills. Only integer or numeric column is allowed.}

\item{time_id}{a string, column name which indicates sessions. Only integer or numeric column is allowed.}

\item{vote_col}{a string, column name which indicates the votes. Only integer or numeric column is allowed.}

\item{bill_match}{(Optional). (Optional). An integer vector which indicates matched bills (the location of matched bills, starts from 0)
and is the same length as the number of bills. If a bill does not have matches, please fill NA. Using \link{make_bill_match} is strongly recommended.}

\item{drop_unanimous}{a bool, whether removing unanimous bills or not. Default is FALSE.}
}
\description{
\code{make_dyn_options} is an auxiliary function for dynamic pgIRT model and generates options for such models.
}
