% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dyn_options.R
\name{make_dyn_options}
\alias{make_dyn_options}
\title{Generating options for dynamic pgIRT model}
\usage{
make_dyn_options(
  dataframe,
  unit_id = NULL,
  bill_id = NULL,
  time_id = NULL,
  vote_col = NULL,
  add_bill_match = NULL,
  drop_unanimous = FALSE
)
}
\arguments{
\item{dataframe}{data.frame, tbl_df or matrix object; dataframe (long-format, i.e. voter-bill unit) of voting data.}

\item{unit_id}{string, column name which indicates voters.}

\item{bill_id}{string, column name which indicates bills}

\item{time_id}{string, column name which indicates times.}

\item{vote_col}{string, column name which indicates the votes.}

\item{add_bill_match}{Optional. Numeric vector which indicates matched votes.}

\item{drop_unanimous}{bool, whether removing unanimous bills or not}
}
\description{
\code{make_dyn_options} is an auxilary function for dynamic pgIRT model and generates options for such models.
}
