% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_init.R
\name{make_init}
\alias{make_init}
\title{Generating initial values for pgIRT model}
\usage{
make_init(
  data,
  model = c("bin", "bin_dyn", "multi", "multi_dyn"),
  T = NULL,
  bill_match = NULL,
  constraint = NULL
)
}
\arguments{
\item{data}{a matrix object, roll-call matrix}

\item{model}{string, one of "bin" (binary), "bin_dyn" (dynamic binary), "multi" (multinomial) or "multi_dyn" (dynamic multinomial)}

\item{T}{string, indicating the length of times. If you choose "bin_dyn" or "multi_dyn" for \code{model}, you must supply this argument.}

\item{bill_match}{Numeric vector which indicates matched votes.}

\item{constraint}{a double, indicating the voter whose ideal point is always set positive}
}
\description{
\code{make_init} is an auxilary function for pgIRT model and generates optimal initial values based on roll-call matrix.
}
