% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rollcall.R
\name{make_rollcall}
\alias{make_rollcall}
\title{Generating roll-call matrix for pgIRT model}
\usage{
make_rollcall(
  dataframe,
  unit_id = NULL,
  bill_id = NULL,
  vote_col = NULL,
  drop_unanimous = FALSE
)
}
\arguments{
\item{unit_id}{string, column name which indicates voters.}

\item{bill_id}{string, column name which indicates bills}

\item{vote_col}{string, column name which indicates the votes.'}

\item{drop_unanimous}{bool, whether removing unanimous bills or not}

\item{@param}{dataframe data.frame, tbl_df or matrix object; dataframe (long-format, i.e. voter-bill unit) of voting data.}
}
\description{
\code{make_rolllcall} is an auxilary function for pgIRT model
and generates roll-call matrix from long-dataframe.
}
