% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgIRT-package.R
\docType{package}
\name{pgIRT-package}
\alias{pgIRT-package}
\title{pgIRT: Fast IRT estimation with Polya-Gamma data augmentation and EM Algorithm}
\description{
This package implements fast IRT model with Polya-Gamma data augmentation and EM Algorithm.
}
\keyword{internal}
