% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgIRT.R
\name{pgIRT}
\alias{pgIRT}
\title{Polya-Gamma IRT with EM algorithm}
\usage{
pgIRT(
  data,
  prior = NULL,
  init = NULL,
  constraint,
  model = c("bin", "bin_dyn", "multi", "multi_dyn"),
  dyn_options = NULL,
  tol = 1e-06,
  maxit = 500,
  verbose = NULL
)
}
\arguments{
\item{data}{a matrix, roll-call matrix. For binary model, the matrix is allowed to contain only 1, 0 and NA. For multinomial, 1, 2, 3 and NA is allowed.}

\item{prior}{a list, containing prior distribution.}

\item{init}{a list, containing initial values.}

\item{constraint}{a double, indicating the voter whose ideal point is always set positive}

\item{model}{string, one of "bin" (binary), "bin_dyn" (dynamic binary), "multi" (multinomial) or "multi_dyn" (dynamic multinomial)}

\item{dyn_options, }{a list, containing the options for dynamic model. If you choose "bin_dyn" or "multi_dyn" for \code{model}, you must supply this argument.}

\item{tol}{a double, convergence threshold. Default is 1e-6.}

\item{maxit}{a double, maximum number of iterations. Default is 500.}

\item{verbose}{a double, the function prints the status every \code{verbose}.}
}
\description{
\link{pgIRT} Estimates IRT model (binary, multinomial and dynamic) by Polya-Gamma data augmentation and EM algorithm.
}
