% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgIRT_boot.R
\name{pgIRT_boot}
\alias{pgIRT_boot}
\title{Parametric bootstrap for pgIRT}
\usage{
pgIRT_boot(fit, boot = 100, verbose = NULL)
}
\arguments{
\item{fit}{pgIRT object from \code{pgIRT()}.}

\item{boot}{an integer, number of bootstrap iterations.}

\item{verbose}{an integer, the function prints the status every \code{verbose}.}
}
\description{
\link{pgIRT_boot} implements paramtric bootstrap for pgIRT model to compute statistical uncertainty.
}
