% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.pgIRT.R
\name{print.summary.pgIRT_fit}
\alias{print.summary.pgIRT_fit}
\title{Print function for EM}
\usage{
\method{print}{summary.pgIRT_fit}(object)
}
\arguments{
\item{object}{An output of \code{summary.pgIRT_fit()} function.}
}
\description{
Print function for EM
}
