% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.pgIRT.R
\name{summary.pgIRT_fit}
\alias{summary.pgIRT_fit}
\title{Summary function for EM}
\usage{
\method{summary}{pgIRT_fit}(fit, parameter = c("alpha", "beta", "theta"))
}
\arguments{
\item{fit}{an object class of \code{"pgIRT_fit"}}

\item{parameter}{a string or string vector, indicating what parameters you want to get.}
}
\description{
Summary function for EM
}
